streamlit-apps/personal_rag_assistant/requirements.txt
streamlit==1.50.0
langchain==1.0.8
langchain-community==0.4.1
chromadb==1.4.0
pypdf==6.5.0
streamlit-authenticator==0.4.2
pyyaml==6.0.3